requirements.txt
pydantic
PyYAML
loguru
jinja2
typer
requests
fastapi
uvicorn
nvidia-riva-client==2.25.0
grpcio==1.67.1
grpcio-status==1.67.1
protobuf==5.28.3
